requirements.txt
autopep8==1.3.2
backports-abc==0.5
BaseHash==3.0.4
certifi==2017.7.27.1
chardet==3.0.4
django>=1.11.28
django-silk==1.0.0
djangorestframework==3.9.1
factory-boy==2.9.2
Faker==0.7.18
gprof2dot==2016.10.13
idna==2.6
ipaddress==1.0.18
Jinja2>=2.10.1
MarkupSafe==1.0
psycopg2==2.7.3
pycodestyle==2.3.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests>=2.20.0
singledispatch==3.4.0.3
six==1.10.0
snakeviz==0.4.2
sqlparse==0.2.3
tornado==4.5.1
urllib3>=1.24.2
MySQL-python==1.2.5